fix for assoc
I made a couple more changes to it. Columns should have a backtick ( ` )...I think.
Also if the query returns false it will go directly to else, we don't have to do separate error handling for it and if it returns anything else, only then $row will be defined. Hope this gets it to work.