Home Page (Final)(Tested) (NO CSS)
So the page works, CSS for error has a class name error it's in a span so change class name if you want but not any form elements. So that's that. It will start a session and let us carry the vehicle number to the next page we're on. (Add The civilian dash page )