Civilian View File. Redirect after Vno is input (UNTESTED)
The Logout button is NOT a hyperlink and instead calls a PHP function that ends the session and then redirects you.
I have added the display part. It will create a new row of data after every iteration instead of one. This way we're covered for multiple offences. @Abbaskt had already added a class. If you want to add more just build CSS and tell me what the class should apply to and I'll update it. I'll start building readme in the morning (5/5:30) if it's incomplete text me on AAJA.